ENH: use separate vars for creating c++ and c shared libraries and add a test for c libraries

+++ Source/cmExecProgramCommand.cxx
@@ -28,7 +28,7 @@ bool cmExecProgramCommand::InitialPass(std::vector<std::string> const& args)
   std::string arguments;
   bool doingargs = false;
   int count = 0;
-  for(int i=0; i < args.size(); ++i)
+  for(size_t i=0; i < args.size(); ++i)
     {
     if(doingargs)
       {
